I think this is enough test cases (techinically ~104k at the minimum, were doing the dataprovider thing ourselves to avoid killing TestNG)

+++ java-manta-client-unshaded/src/main/java/com/joyent/manta/util/MultipartInputStream.java
@@ -67,7 +67,7 @@ public class MultipartInputStream extends InputStream {
     /**
      * Total count of bytes read across all wrapped streams.
      */
-    private long count;
+    private int count;
 
     /**
      * Creates a new instance with the default buffer size.
@@ -201,7 +201,7 @@ public class MultipartInputStream extends InputStream {
         this.closed = true;
     }
 
-    public long getCount() {
+    public int getCount() {
         return this.count;
     }
